+++ CLEYERA/Cleyera.h
@@ -62,7 +62,7 @@ public:
 	/// <param name="Resource"></param>
 	static void TriangleDraw(
         Vector3 position,
-		int size,
+		float size,
 		unsigned int Color,
 		WorldTransform worldTransform,
 		ResourcePeroperty Resource);
